requirements.txt
PyQt6==6.7.0
pillow==10.3.0